HACK4DELHI/hack4delhi_backend/requirements.txt
numpy 
langchain
langchain-core
langchain-community
pypdf
pymupdf
langchain_text_splitters
sentence-transformers
chromadb
langchain-groq
langchain-google-genai
Flask==3.1.2
flask-cors==6.0.2
trafilatura==2.0.0
langdetect==1.0.9
youtube-transcript-api==1.2.3
python-dotenv==1.2.1
requests==2.32.5
transformers==4.57.3